movie-trailer-app-no-docker/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
